Fixed bugs in walking a stack that includes a trap frame. The stack frame walker is now aware of trap frames when checking for an illegal call from a Java frame into a @C_FUNCTION annotated frame.
Fixed bug in assembler/disassembler to do with decoding of inline data.

Turned off stack trace creation for OutOfMemoryErrors.

Made stack walker ignore trap frames when doing a GC while handling a stack overflow error.
Such a GC is not unlikely when trying to fill in the stack trace for a StackOverflowError which can
obviously contain thousands of frames. The trap frame in this situation has no object references
that need to survive the GC.