Added new -XX:PrintMetrics VM option to report all global metrics gathered in VM run. This includes a few new timers for the various IR translators.
Added new annotation @RESET annotation to be used for fields whose value should be reset to the default value of the appropriate kind when copied into the boot image.

Created thread specifics for primordial thread to enable detection of traps on this thread (which should not occur).

Extended VMOption and VMOptions to support the notion of sub-options (e.g. -verbose / -verbose:gc / -verbose:class ...).

Added hook for a VMOption to perform some action (such as printing statistics) just before the VM exits.