Added CFGPrinter class to produce control flow graphs in the format consumed by the C1 Visualizer (https://c1visualizer.dev.java.net/).
Added methods to Util.java for formatting a CiType, CiMethod and CiField to a string.

Renamed C1XPrintStream to LogStream. This stream now maintain a line buffer (in a StringBuilder instance) which is flushed every time a newline is written to it. It now also supports indentation as required by the CFGPrinter class. Lastly, each of its 'print()' and 'println()' methods return the LogStream instance on which they were invoked. This allows chaining of these calls to mitigate use of String concatenation by the caller.

Added a TTY class to provide the same (or similar) logging functionality provided by the 'tty' global variable in HotSpot. When this class is built into the boot image, its underlying output stream is replaced with Log.os object from the VM project.

Extended C1XTest so that command line options are automatically generated for each boolean option defined in C1XOptions. The syntax of these command line options closely mirrors that used by the HotSpot -XX style boolean options. For example, -XX:+PrintCFGToFile.