Made C1XTargetMethods record which calls are to native code.
Added mechanism for denoting a platform register to be non-zero any time it is read by compiled code (e.g. the thread-locals register %r14 on AMD64).

Completed C1X implementation of atomic builtins.

Changed mechanism for traversing all/some values ina FrameState from one that built a new collection for each traversal to one that uses a call back for each value traversed (removing the overhead of the collections).

Fixed bug in LSRA algorithm in the optimization to eliminate redundant spill stores.

Made it legal for an LIR instruction to have a result operand initially bound to a register in the case where the register is not available to the register allocator and is guaranteed never to store a reference.

Split some LSRA-related mechanisms into two separate code paths: one for the -C1X:+PrintCFGToFile option and one for console tracing.

Made -C1X:-GenSpecialDivChecks false by default as Maxine's native trap handler can determine if the special case of Integer.MAX_VALUE / -1 caused the trap.