Fixed side-effect problem at boot image generation time that caused prohibited package to be pulled in the prototype whenusing alternate heap scheme. The problem is that using schemes in the host that are different from schemes used in the target causes static initializers specific to host-scheme to be run, which in turn may cause global variable used as root of boot image generation to include class specific to the host. In particular, the commonly used idiom of registering vm Options at static initialization time cause these to be added to a global list used later on at boot generation time.
The difference in non-platform schemes between host and target originates from the early creation of the prototype's host (via MaxineVM.host()). In the absence of any information, this one uses "default" scheme. Changing the default HeapScheme from SemiSpace to Prototype does solves this issues as the later is bland and doesn't include an Options nor-perform any side-effect that impact image generation.

This is just work-around to the more general problem of having less than perfect isolation between target and host.