Further refine HeapScheme API to: - specify whether the boot image loader should reserve some virtual address space and how much, and store the result in   Heap.reservedVirtualSpace - specify whether the boot region should be mapped in the reserved virtual space (start or end) or whether it can be mapped anywhere
Also moved interface to create code manager to the HeapScheme. A heap scheme may need to enforce some address ordering of the code regions
(with respect to other heap regions) and may need to do implementation specific management of code, hence the code manager is best
supplied by the HeapScheme.