Fixed bootstrapping issues revealed by the signal dispatcher thread. The thread was performing indirectly some heap allocation. This sometimes work, and sometimes not (flagrant with the MS heaps scheme). The reason was that the thread was racing with the main thread doing pristine initialization of the schemes, and in particular, the heap scheme. That's where heap initialization (e.g., virtual memory allocation, formatting of heap data structure, etc...) takes place.
In general, it is more robust to bring the VM to the pristine state, where most of the important VM state is initialized, as early as possible. Especially as the VM increase the number of "system threads".

Also fixed some inconsistency of vm's phase management (e.g., the
SpecialReferenceManager was passed the STARTING phase from
place where the VM was in PRISTINE phase.