Generalized safepoint intrinsic/builtin to be an 'infopoint'. These are locations for which debug info is recorded. There are 3 flavors (so far):
SAFEPOINT: emits the instruction(s) that enable a thread to be safely stopped for a VM operation (e.g. a GC).
HERE: emits an instruction that loads it's own address. This is useful when initiating a stack walk from the current execution position (e.g. building the stack trace for an exception).
INFO: does not emit any instructions. This is useful for ensuring that debug info is available for the next instruction (e.g. a guardpoint in Maxpath).

This generalization is supported by both the CPS and C1X compilers.