Uniformized reservation of virtual space to back up code cache across all heap scheme. This exploits the scheme introduced for the MS and MSHeapScheme wherein a HeapScheme can specified to boot image loader to reserve a contiguous range of virtual space of a specific size, and optionally, to map the boot region at a particular location in that space (e.g., start, end, anywhere). This is now used by default in conjunction with the (poorly named) FixedAddressCodeManager to enforce that boot code region is adjacent to the code region allocated at startup.
Basically, the HeapSchemeAdaptor returns a FixedAddressCodeManager that always map its one region immediately after the boot heap region. The HeapSchemeAdaptor specify the boot image loader to reserve 1 G of space and to map the boot region at the start of that space. The FixedCodeManager can now allocate up to 1 G (minus boot code region size) of space for the dynamic code region. This eliminate the need to allocate the boot heap region in the 0-2^32-1 range as was done until now.

We may consider removing the code in VirtualMemory that supported that special case, as well as other code managers since they aren't used anymore by any configuration.