Terminology change. All points in VM managed compiled (or manually assembled) code for which debug information is available are now called "safepoints" instead of "stops". Call instructions are now simply a special case of a safepoint. The platform-specific instruction sequence emitted to trigger a safepoint trap is now called a "safepoint poll". The set of safepoints associated with an instance of code are encapsulated in the new Safepoints class.
Separated the logic behind the -XX:DeoptimizeALot=<n> option out of the Deoptimization class. Now two separate operations are performed to implement this. This first selects a bunch of live methods, the second deoptimizes them. This substantially simplified the core deopt implementation.

Moved test.output.PhyMemTest to demo.PhyMemTest as it is not really an output test.