Added support in T1X for templates with object literals. Scalar literals are still unsupported. This change in combination with removing the @NEVER_INLINE annotation from T1XRuntime means that some T1X templates (such as object allocation and monitor operations) now include inline fast paths.
Removed interface for instruction relative addressing with a displacement relative to the start of the instruction.

Made stack dumping during a FatalError a little more robust.

Added output for the machine code of a MaxTargetMethod when -C1X:+PrintCFGToFile is enabled. This enables one to see the code at its relocated address with all data patches fixed up.