deopt: fixed erroneous initialization of vtable/itable with compiled that caused corrupted vtable in case of deopt.
A deoptimization may occur between the initialization of the vtable/itable of a class actor being defined and the insertion of the class actor to the class hierarchy graph.
In this interval of time, the class actor is not visible to the deoptimization operation. As a result, the vtable/itable of the ClassActor being defined are not reset to vtable/itable trampoline, but may now point to obsolete target methods whose entry points have been patched with a jump to the static trampoline.This causes subsequent vtable / itable dispatch to fail as these results in taking the trampoline: this one attempt to patch what it believes to be a direct call, which results in an fatal error because the call site isn't a direct call but a register indirect call of a vtable dispatch.

The fix is to delay initializing the vtable/itable with compiled methods, and only initialize them with vtable/itable trampoline.The vtable/itable can be refreshed with compiled method entry points once the ClassActor is entered in the lass hierarchy.This delay has the double advantage of avoiding the bug above, and avoid looking up compiled code if the ClassActor being constructed end up not being selected for class definition (because of concurrent class loading for the same class).